Increase subway turnstile animation time
It seems that whoever implemented this reimplementation had a faster computer than I. With the shorter turnstyle time, I fail to make it and have to re-retrieve the token about three times out of four.

This isn't supposed to be a twitch game, but a puzzle game. Redoing the same parts over isn't useful, so I've increased the time to make it hard NOT to make it.